refactor(spelling): extract remote action handling
